formatted code and added comment to api sections

+++ BACKEND/Classes/Task.php
@@ -16,6 +16,8 @@ class Task extends TaskModel
     private $status;
     private $user;
     private $conn;
+    private $taskExist = true;
+
 
     function __construct($userEmail)
     {
@@ -56,6 +58,12 @@ class Task extends TaskModel
     function getOneTask($id)
     {
         try {
+            //check if task exists
+            $obj = $this->taskExist($id);
+            if (!$obj) {
+                http_response_code(404);
+                return die("File not Found");
+            }
             $query = "SELECT title,description,dateCreated,dueDate,status FROM `task_table` WHERE userEmail ='" . $this->user . "'AND id = $id ;";
             $stmt = $this->conn->prepare($query);
             $stmt->execute();
@@ -74,10 +82,19 @@ class Task extends TaskModel
         $alltask = $stmt->fetchAll(PDO::FETCH_OBJ);
         return json_encode($alltask);
     }
+
     public function updateTask($id, array $data)
     {
 
         try {
+
+            //check if task exists
+            $obj = $this->taskExist($id);
+
+            if (!$obj) {
+                http_response_code(404);
+                return die("File not Found");
+            }
             //prepare query
             $params = "";
 
@@ -108,6 +125,12 @@ class Task extends TaskModel
     {
 
         try {
+            //check if task exists
+            $obj = $this->taskExist($id);
+            if (!$obj) {
+                http_response_code(404);
+                return die("File not Found");
+            }
             //prepare query
             $query = "DELETE FROM `task_table` WHERE userEmail ='" . $this->user . "'AND id = $id ;";
 
@@ -129,7 +152,7 @@ class Task extends TaskModel
         foreach ($data as $value) {
             $ids = $ids . " $value,";
         };
-         $ids = substr($ids,0,-1);
+        $ids = substr($ids, 0, -1);
         try {
             //prepare query
             $query = "DELETE FROM `task_table` WHERE userEmail ='" . $this->user . "'AND id IN ($ids) ;";
@@ -143,6 +166,21 @@ class Task extends TaskModel
         } catch (PDOException $e) {
             http_response_code(501);
             echo $e;
+        }
+    }
+    public function taskExist($id)
+    {
+
+        $query = "SELECT * FROM  `task_table` WHERE userEmail ='" . $this->user . " AND id = $id'; ";
+        $stmt = $this->conn->prepare($query);
+        $stmt->execute();
+        $result = $stmt->fetch(PDO::FETCH_OBJ);
+        if (!$result) {
+            $this->taskExist = false;
+            return false;
+        } else {
+            $this->taskExist = true;
+            return true;
         }
     }
 }
